twitter/functions/twitter_compute_load_graph/requirements.txt
elasticsearch>=7.0.0,<8.0.0
elasticsearch-dsl>=7.0.0,<8.0.0
google-cloud-secret-manager==2.2.0
google-cloud-firestore==1.7.0
google-cloud-pubsub==1.5.0
azure-cosmosdb-table==1.0.6
neo4j==4.2.1
furl==2.1.0
